msme-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.15.0
Pillow==10.0.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0